requirements.txt
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==1.0.1
psycopg2~=2.9.3
tensorflow-cpu~=2.8.1
transformers~=4.19.2